create RestrictedButton.swift
